Move build exception to specialised package.
